Publish RockBot.A2A.Gateway as a NuGet library (#279)
Splits the gateway into a packable library exposing ASP.NET Core
extensions (AddA2AHttpGateway, AddA2AApiKeyAuthentication,
MapA2AHttpGateway) and a thin RockBot.A2A.Gateway.Host executable that
owns appsettings.json and the container entrypoint. The library no
longer references RockBot.Messaging.RabbitMQ — it only needs the
messaging abstractions; RabbitMQ moves to the Host project.
