+++ deploy/Dockerfile.a2a-gateway
@@ -11,7 +11,7 @@ COPY src/ src/
 COPY tests/ tests/
 
 RUN --mount=type=cache,target=/root/.nuget/packages \
-    dotnet publish src/RockBot.A2A.Gateway/RockBot.A2A.Gateway.csproj \
+    dotnet publish src/RockBot.A2A.Gateway.Host/RockBot.A2A.Gateway.Host.csproj \
     -c Release -o /app/publish
 
 # Runtime stage — ASP.NET Core needed for web hosting
@@ -26,4 +26,4 @@ RUN chown -R rockbot:rockbot /app
 USER rockbot
 
 EXPOSE 5200
-ENTRYPOINT ["dotnet", "RockBot.A2A.Gateway.dll"]
+ENTRYPOINT ["dotnet", "RockBot.A2A.Gateway.Host.dll"]
