+++ src/RockBot.A2A.Gateway.Host/appsettings.json
@@ -1,7 +1,7 @@
 {
   "Gateway": {
     "AgentName": "RockBot",
-    "Description": "Personal AI agent \u2014 accepts notifications and availability queries",
+    "Description": "Personal AI agent — accepts notifications and availability queries",
     "Version": "1.0",
     "TaskStorePath": "tasks.json",
     "PushNotificationConfigStorePath": "push-configs.json",
