+++ src/RockBot.A2A.Gateway/RockBot.A2A.Gateway.csproj
@@ -1,13 +1,13 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
+    <OutputType>Library</OutputType>
     <TargetFramework>net10.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <RootNamespace>RockBot.A2A.Gateway</RootNamespace>
-    <IsPackable>false</IsPackable>
-    <UserSecretsId>rockbot-a2a-gateway</UserSecretsId>
+    <IsPackable>true</IsPackable>
+    <Description>HTTP A2A gateway primitives for RockBot: ASP.NET Core extensions that expose a JSON-RPC A2A endpoint and agent-card discovery, bridging HTTP requests to a RockBot agent over the messaging bus.</Description>
   </PropertyGroup>
 
   <ItemGroup>
@@ -21,7 +21,6 @@
   <ItemGroup>
     <ProjectReference Include="..\RockBot.A2A.Abstractions\RockBot.A2A.Abstractions.csproj" />
     <ProjectReference Include="..\RockBot.Messaging.Abstractions\RockBot.Messaging.Abstractions.csproj" />
-    <ProjectReference Include="..\RockBot.Messaging.RabbitMQ\RockBot.Messaging.RabbitMQ.csproj" />
   </ItemGroup>
 
 </Project>
