+++ tests/RockBot.A2A.Gateway.Tests/RockBot.A2A.Gateway.Tests.csproj
@@ -17,6 +17,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\RockBot.A2A.Gateway\RockBot.A2A.Gateway.csproj" />
+    <ProjectReference Include="..\..\src\RockBot.A2A.Gateway.Host\RockBot.A2A.Gateway.Host.csproj" />
   </ItemGroup>
 
 </Project>
